Fix archivePathToZip cache and FoundFile
The old implementation had multiple issues. First, the reference
counting was exposed to the user, and surprise, it was not used
correctly. Second, with correct reference counting, the cache did not
evict unused but instead returned also already closed resources. And
last, even if both issues are fixed, the reference count needs to be
performed with the cache eviction, otherwise one might acquire the
resource in between the decrement and eviction.

This patch tries to keep track of shared resources transparently. Each
retrieval from the cache increases the ref count and each close on the
shared resource decreases the ref count. Assuming the user uses
try-resources, this should fix all the issues above.